+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * added the `InstagramTrait`, inheriting methods from the `Instagram` utility
     (which continues to exist, using the trait itself);
 * added the `InstagramComponent`;
+* `PhotosWidgetsCell` uses the `__construct()` method to set an `Instagram` instance;
 * updated for MeCms 2.18.2.
 
 ## 1.4 branch